// src/types/PoolLens.ts
export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface PoolLens extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): PoolLens;
  clone(): PoolLens;
  methods: {
    directory(): NonPayableTransactionObject<string>;

    getPoolAssetsWithData(
      comptroller: string
    ): NonPayableTransactionObject<
      [
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        string,
        boolean,
        string,
        string,
        string,
        string,
        string,
        string,
        string
      ][]
    >;

    getPoolOwnership(comptroller: string): NonPayableTransactionObject<{
      0: string;
      1: boolean;
      2: boolean;
      3: [string, string, boolean, boolean][];
    }>;

    getPoolSummary(comptroller: string): NonPayableTransactionObject<{
      0: string;
      1: string;
      2: string[];
      3: string[];
    }>;

    getPoolUserSummary(
      comptroller: string,
      account: string
    ): NonPayableTransactionObject<{
      0: string;
      1: string;
    }>;

    "getPoolUsersWithData(address[],uint256)"(
      comptrollers: string[],
      maxHealth: number | string | BN
    ): NonPayableTransactionObject<{
      0: [
        string,
        string,
        string,
        string,
        [
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          boolean,
          string,
          string,
          string,
          string,
          string,
          string,
          string
        ][]
      ][][];
      1: string[];
      2: string[];
    }>;

    "getPoolUsersWithData(address,uint256)"(
      comptroller: string,
      maxHealth: number | string | BN
    ): NonPayableTransactionObject<{
      0: [
        string,
        string,
        string,
        string,
        [
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          boolean,
          string,
          string,
          string,
          string,
          string,
          string,
          string
        ][]
      ][];
      1: string;
      2: string;
    }>;

    getPoolsByAccountWithData(account: string): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
      2: string[];
      3: string[];
      4: string[][];
      5: string[][];
      6: boolean[];
    }>;

    getPoolsBySupplier(account: string): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
    }>;

    getPoolsBySupplierWithData(account: string): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
      2: string[];
      3: string[];
      4: string[][];
      5: string[][];
      6: boolean[];
    }>;

    getPublicPoolUsersWithData(
      maxHealth: number | string | BN
    ): NonPayableTransactionObject<{
      0: string[];
      1: [
        string,
        string,
        string,
        string,
        [
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          string,
          boolean,
          string,
          string,
          string,
          string,
          string,
          string,
          string
        ][]
      ][][];
      2: string[];
      3: string[];
      4: boolean[];
    }>;

    getPublicPoolsWithData(): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
      2: string[];
      3: string[];
      4: string[][];
      5: string[][];
      6: boolean[];
    }>;

    getUserSummary(account: string): NonPayableTransactionObject<{
      0: string;
      1: string;
      2: boolean;
    }>;

    getWhitelistedPoolsByAccount(account: string): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
    }>;

    getWhitelistedPoolsByAccountWithData(
      account: string
    ): NonPayableTransactionObject<{
      0: string[];
      1: [string, string, string, string, string][];
      2: string[];
      3: string[];
      4: string[][];
      5: string[][];
      6: boolean[];
    }>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}